+++ mayan/apps/events/icons.py
@@ -3,6 +3,9 @@ from mayan.apps.appearance.classes import Icon
 icon_current_user_events = Icon(
     driver_name='fontawesome', symbol='file-export'
 )
+icon_current_user_events_clear = Icon(
+    driver_name='fontawesome', symbol='broom'
+)
 icon_current_user_events_export = Icon(
     driver_name='fontawesome', symbol='file-export'
 )
@@ -10,10 +13,16 @@ icon_event_types_subscriptions_list = Icon(
     driver_name='fontawesome', symbol='list-ol'
 )
 icon_events_for_object = Icon(driver_name='fontawesome', symbol='list-ol')
+icon_events_for_object_clear = Icon(
+    driver_name='fontawesome', symbol='broom'
+)
 icon_events_for_object_export = Icon(
     driver_name='fontawesome', symbol='file-export'
 )
 icon_events_list = Icon(driver_name='fontawesome', symbol='list-ol')
+icon_events_list_clear = Icon(
+    driver_name='fontawesome', symbol='broom'
+)
 icon_events_list_export = Icon(
     driver_name='fontawesome', symbol='file-export'
 )