+++ mayan/apps/events/permissions.py
@@ -4,6 +4,9 @@ from mayan.apps.permissions import PermissionNamespace
 
 namespace = PermissionNamespace(label=_('Events'), name='events')
 
+permission_events_clear = namespace.add_permission(
+    label=_('Clear the events of an object'), name='events_clear'
+)
 permission_events_export = namespace.add_permission(
     label=_('Export the events of an object'), name='events_export'
 )