+++ mayan/apps/events/views/clear_views.py
@@ -8,53 +8,58 @@ from mayan.apps.views.generics import ConfirmView
 from mayan.apps.views.mixins import ExternalContentTypeObjectViewMixin
 
 from ..classes import EventType
-from ..permissions import permission_events_export
-from ..tasks import task_event_queryset_export
+from ..permissions import permission_events_clear
+from ..tasks import task_event_queryset_clear
 
 __all__ = (
-    'CurrentUserEventExportView', 'EventListExportView',
-    'ObjectEventExportView', 'VerbEventExportView'
+    'CurrentUserEventClearView', 'EventListClearView',
+    'ObjectEventClearView', 'VerbEventClearView'
 )
 
 
-class EventExportBaseView(ConfirmView):
-    object_permission = permission_events_export
+class EventClearBaseView(ConfirmView):
+    object_permission = permission_events_clear
 
     def get_extra_context(self):
         return {
             'message': _(
-                'The process will be performed in the background. '
-                'The exported events will be available in the downloads area.'
+                'This action is not reversible. The process will be performed '
+                'in the background. '
             )
         }
+
+    def get_task_extra_kwargs(self):
+        return {}
 
     def view_action(self):
         decomposed_queryset = QuerysetParametersSerializer.decompose(
             _model=Action, **self.get_queryset_parameters()
         )
 
-        task_event_queryset_export.apply_async(
-            kwargs={
-                'decomposed_queryset': decomposed_queryset,
-                'user_id': self.request.user.pk
-            }
-        )
+        task_kwargs = {
+            'decomposed_queryset': decomposed_queryset,
+            'user_id': self.request.user.pk
+        }
+
+        task_kwargs.update(self.get_task_extra_kwargs())
+
+        task_event_queryset_clear.apply_async(kwargs=task_kwargs)
 
         messages.success(
             request=self.request, message=_(
-                'Event list export task queued successfully.'
+                'Event list clear task queued successfully.'
             )
         )
 
 
-class EventListExportView(EventExportBaseView):
-    object_permission = permission_events_export
+class EventListClearView(EventClearBaseView):
+    object_permission = permission_events_clear
 
     def get_extra_context(self):
         context = super().get_extra_context()
         context.update(
             {
-                'title': _('Export events'),
+                'title': _('Clear events'),
             }
         )
         return context
@@ -65,15 +70,15 @@ class EventListExportView(EventExportBaseView):
         }
 
 
-class ObjectEventExportView(
-    ExternalContentTypeObjectViewMixin, EventExportBaseView
+class ObjectEventClearView(
+    ExternalContentTypeObjectViewMixin, EventClearBaseView
 ):
     def get_extra_context(self):
         context = super().get_extra_context()
         context.update(
             {
                 'object': self.external_object,
-                'title': _('Export events of: %s') % self.external_object,
+                'title': _('Clear events of: %s') % self.external_object,
             }
         )
         return context
@@ -83,9 +88,15 @@ class ObjectEventExportView(
             '_method_name': 'any', 'obj': self.external_object
         }
 
+    def get_task_extra_kwargs(self):
+        return {
+            'target_content_type_id': self.external_object_content_type.pk,
+            'target_object_id': self.external_object.pk
+        }
 
-class CurrentUserEventExportView(ObjectEventExportView):
-    object_permission = permission_events_export
+
+class CurrentUserEventClearView(ObjectEventClearView):
+    object_permission = permission_events_clear
 
     def get_external_object(self):
         return self.request.user
@@ -95,8 +106,11 @@ class CurrentUserEventExportView(ObjectEventExportView):
             '_method_name': 'actor', 'obj': self.external_object
         }
 
+    def get_task_extra_kwargs(self):
+        return {}
 
-class VerbEventExportView(EventExportBaseView):
+
+class VerbEventClearView(EventClearBaseView):
     def get_extra_context(self):
         context = super().get_extra_context()
         context.update(