+++ mayan/apps/views/mixins.py
@@ -181,8 +181,8 @@ class ExternalContentTypeObjectViewMixin(
     external_object_pk_url_kwarg = 'object_id'
 
     def get_external_object_queryset(self):
-        content_type = self.get_content_type()
-        self.external_object_class = content_type.model_class()
+        self.external_object_content_type = self.get_content_type()
+        self.external_object_class = self.external_object_content_type.model_class()
         return super().get_external_object_queryset()